chore(rspack-provider): using data URI as virtual module
