+++ packages/builder/builder-rspack-provider/src/plugins/swc.ts
@@ -6,6 +6,7 @@ import {
   logger,
   setConfig,
   isWebTarget,
+  addCoreJsEntry,
 } from '@modern-js/builder-shared';
 import type {
   BuilderPlugin,
@@ -22,27 +23,9 @@ export const builderPluginSwc = (): BuilderPlugin => ({
   name: 'builder-plugin-swc',
 
   setup(api) {
-    const polyfillEntryFileName = 'rspack-polyfill.js';
-
-    api.modifyBundlerChain(async (chain, { target }) => {
+    api.modifyBundlerChain(async (chain, { isServer, isServiceWorker }) => {
       const config = api.getNormalizedConfig();
-      const mode = config?.output?.polyfill ?? 'entry';
-      const { entry } = api.context;
-      if (['modern-web', 'web'].includes(target) && mode === 'entry') {
-        Object.keys(entry).forEach(entryName => {
-          chain.entry(entryName).prepend(polyfillEntryFileName);
-        });
-
-        const { default: RspackVirtualModulePlugin } = await import(
-          'rspack-plugin-virtual-module'
-        );
-
-        chain.plugin('rspack-core-js-entry').use(RspackVirtualModulePlugin, [
-          {
-            [polyfillEntryFileName]: `import 'core-js'`,
-          },
-        ]);
-      }
+      addCoreJsEntry({ chain, config, isServer, isServiceWorker });
     });
 
     api.modifyRspackConfig(async (rspackConfig, { target }) => {
