+++ packages/solutions/app-tools/package.json
@@ -95,7 +95,6 @@
     "@modern-js/utils": "workspace:*",
     "es-module-lexer": "^1.1.0",
     "esbuild": "0.15.7",
-    "rspack-plugin-virtual-module": "0.1.0",
     "@swc/helpers": "0.5.1"
   },
   "devDependencies": {
